Design Picker: Refactor themes request to react-query and only fetch free themes (#59342)
* Move <MShotsImage> to @automattic/onboarding package

<MShotsImage> has no dependency on the design picker, yet is forcing
other components (e.g. DIFM and the page pattern modal) to pull in all
the design picker code.

Could have moved <MShotsImage> to either the onboarding or components
package. Decided on the onboarding package because it's CSS depends on
the onboarding CSS.

* useThemeDesignsQuery hook requests themes for the design picker

* Design picker requests themes using react-query

Themes were being requested using the Redux store, however this wasn't
wasn't shareable with other consumers of the design picker.

* Design picker step only requests free themes

* Use network response to correctly set `is_premium` flag

* Hardcoded DIFM themes are now Designs

Since the design picker now provides a hook for fetching themes in the
"Design" format, the hardcoded DIFM themes now also need to be in the
"Design" format.