Adds parameter to opt-out from `isManageSiteFlow`
The DIFM Lite flow deals with both new and existing sites. A user may
choose an existing site for DIFM, press back from checkout and then go
into the new site flow. This satisfies the conditions for setting
`isManageSiteFlow` to `true`. However, when the user then goes through
the new site flow, the `siteId` and `siteSlug` will already be set
causing a breakage - the domains step will be skipped because these
dependencies will already be set. This change adds a new optional
parameter `excludeFromManageSiteFlows` that skips setting `isManageSiteFlow`
to `true`.