Improve loading performance of the all domains page
In the all domains page we do a `GET /all-domains/` query, which
retrieves all the domains for all the sites of a user. So at least the
initial domain list render should be quick, because the `all-domans`
query is pretty fast even when you have many domains.

The weird behavior reported in https://github.com/Automattic/wp-calypso/pull/58980#issuecomment-992244728
was happening because the `createLightSiteDomainObject` method was not
processing the `current_user_is_owner` property - it was only set for
domains that were fully loaded before.

This PR adds this property and removes the `AllSitesDomains` query that
was being done upfront and increased the page's load time.