DIFM Lite: Adds flow for choosing existing sites (#58951)
* Adds support for choosing existing sites

* Send difm data in cart item extra

* Linting and TS fixes

* Skip the manage site flow for DIFM

* Styling fixes

* Adds explainer text to the delete dialog

* Copy fix

* Copy and styling changes

* Minor refactor

* Fixes flow for new users

* Re-enable the check for the delete button

* Copy updates

* Adds parameter to opt-out from `isManageSiteFlow`

The DIFM Lite flow deals with both new and existing sites. A user may
choose an existing site for DIFM, press back from checkout and then go
into the new site flow. This satisfies the conditions for setting
`isManageSiteFlow` to `true`. However, when the user then goes through
the new site flow, the `siteId` and `siteSlug` will already be set
causing a breakage - the domains step will be skipped because these
dependencies will already be set. This change adds a new optional
parameter `excludeFromManageSiteFlows` that skips setting `isManageSiteFlow`
to `true`.

* Fixes TS errors