Improve swipe event onLeave
Currently when the onPointeLeave event is called It does't have the any x or y values. We store the last move event value so that they can used to calculate if we should move to the next slide.

We also added an angle that is higher than 45 as well as reduce the speeded need to register the event.