Trending/recommended themes: Only get siteId when state.ui exists
I was seeing this type of failure in unit tests:
```
  ● #getRecommendedThemes › should return empty themes list for unfetched filter

    TypeError: Cannot read properties of undefined (reading 'selectedSiteId')

       8 |  */
       9 | export default function getSelectedSiteId( state ) {
    > 10 |      return state.ui.selectedSiteId;
         |                      ^
      11 | }
      12 |
```

The unit tests were trying to get a selected site id, to check if the
site was a standlone jetpack site, but since state.ui does not exist,
the getSelectedSiteId function was crashing.

It seems a better fix is to change getSelectedSiteId to
state?.ui?.selectedSiteId, but it's such a core function to calypso I'm
weary of changing it.