Remove duplicate key errors from DesignPicker integration test
The themes in available-designs-config.json are duplicated to have a FSE
and non-FSE version (this is because these themes are "universal" and
support being used as both FSE themes and "classic" themes).

All uses of the <DesignPicker> (other than tests) are now providing
their own `design` prop and not relying on the default prop value. This
means they're doing the work of filtering out these duplicates.

This fix was just ensuring that the tests are also filtering out these
duplicate designs.