Hide universal themes from block templates filter
The block templates filter is used to search for full site editing (FSE) themes. Universal themes would normally enable FSE themes. However, sites with a universal classic theme will not enable FSE, even if another universal theme is selected. Because of this, universal themes should not appear in the block templates search filter if the current site is using a universal classic theme.