Plugins: Display chat info in checkout summary for plugins
The live chat support information was not displayed in the checkout
summary when purchasing individual plugins.

To display the live chat support, the current plan information is
extracted using selectors. The following criteria is required to allow
live chat support:
- Plan type
- Billing period

Also, the live chat support is already displayed as part of the plan
summary so it will only be displayed when not purchasing a plan.