Plugins: Display chat info in checkout summary for plugins (#61259)
* Plugins: Display chat info in checkout summary for plugins

The live chat support information was not displayed in the checkout
summary when purchasing individual plugins.

To display the live chat support, the current plan information is
extracted using selectors. The following criteria is required to allow
live chat support:
- Plan type
- Billing period

Also, the live chat support is already displayed as part of the plan
summary so it will only be displayed when not purchasing a plan.

* Refactor checkout chat support to external component

Extracts the logic and display of the chat support in the
checkout details card to an external component. Leaves the
condition `hasPlanInCart` in parent component so it can be easily
identified that the new logic is applied when the plan features are
not applied.

* Add slice of state to tests that selector does not fail

This is to mock the `getBillingInterval` selector

* Use plan period to allow live chat support

Instead of using the plugin period, the plan period is used to display or
hide live chat support.

This change also removes the slice of mock state added for tests that is
not required anymore after removing the selector that used it.