Rename options param to `requiresGutenframe`
The previous name suggested the error would only happen if the user was booted from the Gutenframe, but it might be the case that a direct navigation to wp-admin is made (i.e if the site is configured to not use the Gutenframe), in this case we might still want the test to "require" the Gutenframe if we're testing an aspect related to it, and fail if the Gutenframe is not present. This name better reflects those semantics.