Search: Send users to the right place for customization post-checkout (#60956)
* Refactor test for search status.

The same test is used in three separate functions so best not to repeat ourselves.

* Update copy for shown after purchasing a Search plan.

Just a minor change to make the message more generic. Using the same message for Atomic & Simple sites.

* Move test for Search above My Home test.

We already has the necessary JSX in place but that code path was never hit. With this change, it should be triggered for any purchase of a Search plan.

* Factor out logic for the primary button.

We update the button title and the target URL based on if the site is a Jetpack site or not. Currently returns a placeholder URL for the “Go to Search Dashboard” button.

* Add getJetpackSearchDashboardUrl selector.

Just wired things up. Will be used to return the URL of the Jetpack Search Dashboard. Currently returns ‘#’.

* Add some initial tests around new selector.

Two tests for getJetpackSearchDashboardUrl.

1. Confirm null when site not loaded.
2. Confirm ‘#’ when we have a Jetpack site.

* Implement getJetpackSearchDashboardUrl.

With tests to confirm expected behaviour.

1. Return null when site not loaded.
2. Return null for Simple sites.
3. Return null if adminUrl is not found.
4. Return default dashboard for versions prior to 10.1.
5. Return Search dashboard for version 10.1 and later.

* Updated tests to use Jest.

* Remove unneeded comment.

* Update URL for older versions of Jetpack.

If not on 10.1 or later, send the user to the following:

/admin.php?page=jetpack#/performance