Update the displaying of compact breadcrumbs.
Before this change, the items property was being updated, not allowing it to go back to original value when resizing to a wider screen size.
After this change, the back link is displayed on the screen but the items property remains untouched allowing it to returns to its original value.