Update the Plugin Details page breadcrumb to be based on the url plugin slug.
Before this change, the breadcrumb was based on the slug from the fullPlugin property. This property is mounted by different sources (WP.org, WP.com, etc) which can cause it receive different values, creating a new breadcrumb for the same page.
After this change, the breadcrumb is based on the slug from the url, which shouldn't change.