Update the displaying of compact breadcrumbs (#61625)
* Update the displaying of compact breadcrumbs.

Before this change, the items property was being updated, not allowing it to go back to original value when resizing to a wider screen size.
After this change, the back link is displayed on the screen but the items property remains untouched allowing it to returns to its original value.

* Update the Plugin Details page breadcrumb to be based on the url plugin slug.

Before this change, the breadcrumb was based on the slug from the fullPlugin property. This property is mounted by different sources (WP.org, WP.com, etc) which can cause it receive different values, creating a new breadcrumb for the same page.
After this change, the breadcrumb is based on the slug from the url, which shouldn't change.