Fix issue with Back button when returning from the Woo upgrade flow
It turns out that `goToNextStep` takes an optional argument of the next
flow name. If this argument isn't provided, it defaults to the current
flow name.

We now explicitly pass `woocommerce-install` to `goToNextStep`. This
lets the signup process know that we've moved between flows so it keeps
the state properly and the Back button will work.

We also tweaked `goToNextStep` so that it waits until the state update
is complete before going to the next step. We need this otherwise it
hangs when we enter the woo flow because it's using stale step data.