Payment Methods: Re-throw errors in fetchStoredCards/deleteStoredCard (#61666)
* Re-throw errors in fetchStoredCards/deleteStoredCard

The Redux action creators `fetchStoredCards` and `deleteStoredCard` are
Thunks, and they return a Promise that resolves to the result of the
respective endpoint request. This is then returned by the Redux
dispatcher that dispatches the Thunk and can be used to chain on
additional behavior.

However, in the case of an endpoint failure, both functions would still
return a resolved Promise because they have their own `catch()`
handlers. This is counterintuitive and can create bugs if the caller
expects the Promise to reject on error (as is the case with deleteStoredCard).

With this change, both functions will reject if there is an error.

Co-authored-by: vitozev <vitozev@gmail.com>

* Update tests to be able to handle a rejecting promise

Co-authored-by: vitozev <vitozev@gmail.com>