Don't display support contact card for non-wpcom themes
When a user is viewing the 'Support' tab for a non-WordPress.com theme, hide the 'Need extra help' card (since wpcom doesn't support third-party themes.

This is done by checking `isWpcomTheme` or if `author` is 'Automattic'.
This covers the case of the **Storefront** theme which isn't a wpcom
theme but it still supported since it's from Automattic.