Update theme search to account for twentytwenty-two theme
Normally entering the word "two" anywhere in the theme search pops up
the "Two Column" suggestions.

This change makes it so any search that matches `twenty.*?two` will only
display the search results and not the suggestion.

We do this by adding an `exclusions` prop to the `KeyedSuggestions`
component. This prop is an array of regexes. If the current search term
matches any of those regexes, then we only search for matches by the
full term rather than splitting the term up and searching by similarity.