Marketing Connections: Ensure Google Photos works with the site ID and slug (#61980)
During some testing we saw that if we went to the
`/marketing/connections` screen with the site ID rather than the site
slug/URL, we would see a connection button for Google Photos regardless
of the connection status. The button was also not in a consistent
format.

This was becaus a check in the code only looked for the path being
`:site` and didn't account for the path having `:siteid`. This changes
fixes that.