Breadcrumb: Update styles based on the Domains breadcrumb (#61995)
* Breadcrumb: Update styles based on the Domains one

*Note:* The behaviour of the existing Breadcrumb component has *not* been
modified. Only difference is that the back icon will not be displayed
when the page is in the root component so there is only one item

Additionally, the component has been refactored to cope with three
different conditions:
1. When it is in the compact view and there are multiple items:
  - Only show Back navigation
2. When it is not in the compact view and there are multiple items:
  - Standard breadcrumb behaviour with navigation for the parents
3. When there is only one item:
  - No back button is displayed

* Fix e2e test. Remove li selector as not used in Back button

* Fix a couple of style issues

* Add the bubble functionality to breadcrumbs

It also adds a bubble to the Plugins root item of the breadcrumb.

* Add a mobileItem label to Breadcrumb to render in compact mode

If passed, `mobileItem` label will be rendered in compact mode. If not
passed, the default "Back" label will be rendered in compact mode.

This commit also updates the docs to include two more examples

* Fix warning: React element must return at least one value

The previous change was not correct as all return statements were under
an if condition

* Remove css important mmodifier and use pseudo-classes instead