Switch from useIsSimpleSeller hook to getSiteOption (#62130)
* Switch from useIsSimpleSeller hook to getSiteOption

The hook isn't necessary. All we need to do is use `getSiteOption` to check if the site intent is `sell` on the WooCommerce landing page (the only place the hook was being used).

* Remove useIsSimpleSeller hook (#62110)

Originally, we were using the hook to check both the site intent _and_ that the site plan _wasn't_ "ecommerce-bundle".

Checking the site plan isn't actually necessary since people with the full ecommerce plan will see the WooCommerce admin and not the landing page.