Only access the selected site options when we know one is selected
The domain management <Settings> component can be momentarily rendered
as the user navigates through the all-sites domain management interface,
even in situations where there is no selected site. This causes an error
that breaks rendering when `selectedSite` is undefined.

By delaying access to `selectedSite` till later in the render, by which
time we know there will be a selected site, then this fixes the error.