Allow env variable consumers to know when an env variable is not set, if desirable
It might make sense to respond differently if an env variable hasn't been passed (is `undefined`). Before this changeset, it was impossible to know when an env variable was undefined as it would always fallback its default values set in the `env-variables` module

Semantically, if an env variable value is not set, it might mean that specific artifact is not present. It could mean a plugin that should not be present the test site used for the E2E test. This is in fact a change that would play better with the table-like logic of `getTestAccountByFeature()`.

For example, let's say we want to select a site for testing Gutenberg and that Coblocks is not relevant. By not relevant, it might mean it's not installed or just fallbacks to the default version in the site. Before, it wasn't possible to communicate that, as `COBLOCKES_EDGE` would always fallback to `false` if not set, and `false` meant CoBlcoks stable and not the absence of CoBlocks.