Stripe connected accounts can access earn regardless of plan
The pages on earn to manage various earn settings are only available to
sites with a paid plan, even though payments will continue to be
processed and the stripe connection will still be in place.

This change allows free sites to use the Manage payments, Donations and
Premium Content so long as they have an existing stripe connection.

Fixes #59937