Update license issuing to accept a site_id (#62926)
* Update license issuing to accept a site_id and automatically assign the license to the site_id

Update issuing a license to allow for a site_id to be passed. This will cause the issued license to be automatically assigned to the site_id that is passed and bypass the step to assign the license.

* Update variable name

Update query param and uses from site_id to siteId

* Check site id ownership when issuing a license

Update the controller to verify that the passed site id is one of the users sites before using it when issuing a license.

* Rename query param

Rename the site_id query param

* Update siteId allowed types

Allow null as a possible type for siteId property when attempting to automatically assign an issued license.