Site Selector: Update the Jetpack site filter (#62955)
* Site Selector: Update the Jetpack site filter

While testing some work on the new Jetpack Social plugin, we discovered
that sites weren't showing in the site selector. There is a filter used
to check that we have an active Jetpack connection and one of either the
Jetpack plugin, or the Jetpack product plugins available.

But this filter is defined differently based on if the site selector is
on initial load, or if it is shown when switching site in the sidebar.

This PR consolidates the check into a helper function that can be used
in both places. At the same time it includes the new Jetpack Social
plugin in the list of valid product plugins.

* Restrict Calypso to just sites with the Jetpack plugin.