Display refund windows individually in checkout
Having a domain and a plan in the cart at the same time would previously cause a conflict. The Pro plan is refundable within 14 days, but a domain is only refundable within 4 days. We resolved this by not specifying the number of refund days (instead just saying "Money back guarantee").

With this change, we display multiple rows of refund window info in the checkout summary – one for every product type.

I have also changed the checkout summary to never show the string "Unlimited customer support via email". This string is used in other places for marketing purposes, but there is no need to distinguish between "Unlimited" and not in this context.