Fix unlocalized links everywhere (#63306)
* Fix unlocalized links everywhere

* Fix failing tests by appending a '/' to the path of each URL.

The localizeUrl() function adds a trailing slash '/' to the path in the URL.
This caused tests to fail instances as localized URLs contained a slash
but the original constant in support.js did not.

The trailing slash was difficult to notice when it appeared before a '#'
character.

* Use localizeUrl in locations that were missed

localizeUrl is used in one of the following scenarios:
- where the constant is used in an anchor element
- prior to passing the constant to a React component

If the component uses localizeUrl internally, e.g. PurchaseButton,
then the constant is unaltered.