Masterbar: #49847 enable My sites link on theme section
When going to Appearance > Themes , and selecting any theme, calypso navigates to the theme section.
This section has a css to hide the Sidebar/menu on the left.
`.theme-default.is-sidebar-overflow .is-section-theme .layout__secondary`
https://d.pr/v/xZpyC2

And the My Sites is an empty button that does nothing. This commit converts the My Sites into a link for this specific section.
https://d.pr/v/0SYH0t

Closes #49847