social-previews: Add missing dep on `@emotion/react`
There's no direct reference to `@emotion/react` in the source of the
package, but calypso-babel-config sets up `@babel/preset-react` to use
that rather than normal React so the built code winds up depending on
it.

Following the lead of `@wordpress/components`, this is added as a
dependency rather than a peer dep.