Jetpack Cloud: Ensure data is available for Social connections
When disconnecting connections on load, or quickly attempting to connect
an account, there would be an error, or the process would silently fail.
This appears to be because the data wasn't available in the store.

This change adds the two query components in use on the
`/marketing/connections` screen, that weren't on the Jetpack Social
equivalent, which fixes the issue.