Invalidate relevant queries when email forwards are removed
Email forwards currently rely on both Redux and react-query. This causes confusion when email forwards are removed, because the mailbox is removed in Redux, but not in react-query.

With this approach, we add a react-query mutation that runs when an email forward is removed and then invalidate the relevant react-query queries to fetch updated data.