eslint-plugin-wpcalypso: Add missing peer dep on eslint-plugin-react
Without the peer dep being declared, it won't be found when using yarn's
p'n'p or pnpm with hoisting disabled.

As it's already documented as optional in README.md, mark it as such in
package.json. And mark the eslint-plugin-react-hooks dep optional too,
as it's only needed under the same conditions as eslint-plugin-react.