Disable tracing in for Sentry on WPCOM/GB
Docs: https//docs.sentry.io/platforms/javascript/performance/instrumentation/automatic-instrumentation.

It might come in handy in the future and we might re-introduce it, but it's not the focus of the integration now.

If we ever re-introduce it, Sentry recommends setting a lower sample rate (it shows 0.2 = 20% in the docs). The current 1.0 never caused any issue because as of this commit, Sentry is only activated for 10% of WPCOM simple users.