Marketplace: Do not show eligibility modal when not needed (#64414)
* Marketplace: Do not show eligibility modal when not needed

The eligibility modal should not be displayed if there are no warnings
or additional messages to show

* Fix condition that displays eligibility modal

The `isEligible` condition was preventing the display of the eligibility
modal for _free_ sites. That `isEligible` flag is not required as either
the `eligibilityHolds` or the `eligibilityWarnings` arrays will be
populated and will display the modal.

If there are no values in those arrays, the modal does not need to be
displayed anymore.