Fix timing issue with the site options step
On slow networks, it was possible for the user to edit the site name or
tagline and then have that value overwritten by the value from the
store.

There was a short window where the form inputs were accessible but the
api call to retrieve the site hadn't finished yet. When the api call
finished, the `useEffect` hook would overwrite the form input values
with the data from the store.

Now we check to make sure the form inputs haven't changed before we
default to the values from the store.