Refactor condition that hides AMP settings for deprecated sites
This PR changes the condition that hides AMP settings for deprecates sites
to account for changes done on REST API side.

I do this to prepare the code for incoming UI changes that need to
display different copy for site with deprecated AMP and for site that has it
not supported (eg. private site).