Improve search on the Country/State dropdown of the Store Address step
The dropdown had all of the options formatted as `COUNTRY -- STATE`.
This meant that the user only got results if they typed in an _exact_
match (including the `--`).

We now use the `onFilterValueChange` handler to customize the filtering
of the options the dropdown displays.

We do this by removing the characters `-`, `/`, and `\` from both the
dropdown option and the value input, then do a lowercase comparison to
see if the value is present in the label.

This means the user can type any part of the Country/State combination
and get valid results.