Update theme forum URL (#64976)
We're shutting down the subforums on WordPress.com, including the themes subforum.
This changes the theme forum URL to go to the main support forums instead
for WordPress.com themes.

Also updating the comment to accurately reflect what the code does, as it
still references premium theme forums which also no longer exist.