The __i18n_text_domain__ constant so it can be used by shared components
When components in packages use the __() function, they don't know what
text domain to use for the second parameter. That's because they're a
shared package; they don't know what "domain" they're going to be used
in!

The way we get around this is to use the __i18n_text_domain__ build-time
constant as the text domain. This allows the package to compile without
errors, but when the app using the component gets built the
__i18n_text_domain__ constant gets replaced with the actual text domain
e.g. 'full-site-editing` or 'default'