Include site-editor as a valid editor pattern
The lastNonEditorRoute Reducer handles state for the most recently visited calypso url path that didn't load the gutenberg block editor (Ex. /home, /themes, etc.).

Unfortunately, the logic didn't consider the /site-editor route to be a valid editor. As a result, it was returning /site-editor as a "lastNonEditorRoute", when in reality, it shouldn't. We amend that issue by including /site-editor as a recognized editor route.