Site Editor: Include site-editor as a valid editor for last non editor route
The lastNonEditorRoute Reducer handles state for the most recently visited calypso url path that didn't load the Gutenberg block editor (Ex. /home, /themes, etc.).

Unfortunately, the logic didn't consider the /site-editor route to be a valid editor. As a result, it was returning /site-editor as a "lastNonEditorRoute", when in reality, it should have been excluded. Downstream logic would then provide incorrect back to dashboard urls within the editor. We patch the issue by including /site-editor as a recognized editor route.