Update back-to-top-button.tsx
HelpCenter (BackToTopButton): attach to scroll events of the parent element

The back-to-top button was trying to automatically determine which is the scroll container. However, in the case where the content is currently loading, there will be no changes in height it would fail to determine the correct container.
This PR ties the scroll container to the parentElement of the button