Dynamically generate link
If the FAQ is already on the page, we want to link to it as an anchor link, not in a separate page. So we check for the ID on the page, and if it exists, we just use an anchor link to jump to it on the page. Otherwise a full URL gets sent as the link and it opens a new window