Consolidate template part tracking events to single file
Tracks events for inserting a new template part and replacing a template part had been in separate files. This commit consolidates them into a single file. The reason is that due to a Gutenberg change, these events run based on the same selector, with some conditional logic that is easier and DRYer to handle together.