Re-instate fuzzy search ordering
Fuzzy search leaves items in the table even when they don't technically
match the search term. This means sorting by relevance is an important
part of the fuzzy search experience. Otherwise you can end up with a
perfect match appearing at the bottom of the list, and a row that has
very little to do with the search term appearing at the top.