Editing Toolkit: Add block-editor preview button url override (#66046)
The block editor preview button was throwing a 404 for draft previews because it would lead from a wordpress dotcom domain to a custom mapped domain. Without proper cookie credentials, the redirect would lead to a preview as a logged out user. This fix overrides the preview redirect url with logmein to handle the cookie credentialing problem.