Silence Typescript errors to do with Intl.Segmenter
We usually target the es2020 type library (see
packages/calypso-typescript-config/ts-package.json) but we can opt in to
the es2022 library with a special comment. It's ok to opt-in to these
types because we're being careful to catch any runtime type errors.

However Intl.Segmenter was only added to the es2022 type library in
version 4.7 of Typescript. So until then we need to work around the type
error using `any` (or define our own types for Intl.Segmenter).