Refactor how we override preview url supplied to client
We update the block editor preview url on dotcom to handle third party cookie issues with custom mapped domains.

Previously, we were changing the url dynamically in the client UI on mouse click. This was a brittle approach because it relied on dom selectors to identify the preview button, which could change without notice. This new approach filters the preview url server-side before the page is loaded, and avoids maintenance issues by not relying on said dom selectors.