Fix incorrect visibility of privacy field in image edit form
The privacy field should be displayed only for VideoPress videos, and
it should not be displayed for images.

We are fixing it by adding condition thatb prevents from displaying
the field if item is not a VideoPress video. We also extract the check
to separate method and add tests for both states.