Memo rows to prevent unnecessary renders
The site data objects are treated as immutable. That means if any of the
site attributes changes, that value wouldn't simply be updated in place,
but rather an entirely new site object would be created.

We can use this to memoize the SiteTableRows. If the reference to the
site object hasn't changed it means we can skip rendering the row.