Remove encoding of ampersands in media library edit form fields
The media library item is saved as a custom post type and the
field values are saved with a HTML entities encoded. Then, those
encoded characters are displayed in the media library edit form.

We are changing that behavior to decode those characters before
we display them in the form.