Update the remove action in plugin management
Updating the remove option for bulk actions to no longer disable the button if Jetpack is in the list of selected plugins. A warning will be provided instead once other selected plugins have been removed.

Also disable the remove button similar to existing actions with activate/deactivate when Jetpack is the only plugin selected in the edit all view.