Change Remove subscription button wording in /earn
The Remove Subscription button on the earn page is confusing for
subscriptions which were purchased by one-off donations & payments. The
user is able to access paywalled areas of the site associated with that
payment plan, but don't have an ongoing financial commitment.

The button will now be labelled 'Remove' for one-off payments and
'Cancel Subscription' for ongoing payments.