The Sites Management Page takes over the `/sites` URL
The existing UI at `/sites` looks identical to `/home`, so if the
`build/sites-dashboard` feature flag is disabled then we can just
redirect those users to `/home`.

The section "group" and module are still called "sites-dashboard"
to disambiguate it from the main Calypso dashboard, which is the "sites"
group. That means that CSS rules like `body.is-group-sites-dashboard`
stay as they currently are.