Fix Auto renew status for unbillable payment method
Description:

On the individual purchase management page when the initial payment was
done with a payment method that does not support renewal like ideal,
then we see a message `None` this could be replaced with a text that
makes more sense like `You don't have a payment gateway to renew this
subscription`. Similarly the subscription renewal component before that
says `You will be billed on` without any date information.

Screenshot:

https://d.pr/i/RL3ea9

Suggested Solution:

https://d.pr/i/Bc2I2Y

Testing instructions:

1. Add a renewable addon product to cart like `Removable ads` from /add-ons/:site
2. Ensure to have currency as Euro and country as Netherlands
3. This will provide a gateway like Ideal which will not allow renewal
4. Complete Payment with ideal gateway
5. Now visit /purchases/subscriptions/:site/
6. Click on the exact purchase from active upgrades tab
7. Now, you should see the above suggestion under the `Subscription
   renewal` and `Payment method` columns

Fixes #1058