Create separate banners for the WP and JP apps
With this commit, two separate banners are created for both the WP and JP apps. The JP banner will be displayed when the 'jetpack/app-branding' feature flag is set to true (currently only for development builds and wpcalypso).

With this initial commit, both banners are the same as the current WP banner. The JP banner will be tweaked with future commits.