Update/plugin management remove updates (#66677)
* Update the remove action in plugin management

Updating the remove option for bulk actions to no longer disable the button if Jetpack is in the list of selected plugins. A warning will be provided instead once other selected plugins have been removed.

Also disable the remove button similar to existing actions with activate/deactivate when Jetpack is the only plugin selected in the edit all view.

* Update edit all remove

Update the edit all remove to no longer hide the remove option if Jetpack is the only plugin selected. Also update the remove option to display a warning notice when Jetpack is being removed either on it's own or as part of a bulk action

* Update index.jsx

Update the remove text on the Jetpack warning notice

* Update index.jsx

Switch foreach when setting the isJetpackIncluded value to be more concise