Replace hook with HOC, for easier usage in class
To determine whether an RTL locale is in use, we're currently using the 'useRtl' hook. By replacing the hook with the 'withRtl' HOC, we're able to determine whether a locale is RTL directly within the class component, enabling 'isRTL' to be more cleanly passed down as a prop from the main banner component.

This commit removes the call to the 'useRtl' hook from BannerIcon (child) component. Its 'icon' prop will be defined in future commits.