Newsletter Accent Color: Fix displaying of validity check for the color input
The FormInput component implements css ease-in-out transition for all transition properties, like so:
`transition: all 0.15s ease-in-out;`
As a result when the input value becomes valid we add another background image, the green check, to the current one of swatch only. Therefore at first it transitions from swatch background image to green check. I resolved this by setting the `transition` property to `none` for the input.