Pin new site button to the bottom of the site switcher in JP Cloud
Removing the `display: block` rule was enough to move it to the bottom
(previously it was a cloud-only override to prevent it from being at the
bottom like Calypso). However there was a mysterious extra margin below
the button. I couldn't figure out where this was coming from.

The main JP cloud sidebar uses `margin-top: auto` to pin the `wp-admin`
link to the bottom, so I did the same in the site switcher. That was
enough to remove the margin.