feat: My Sites mobile CTA points to Jetpack app
We want to direct users to install the Jetpack app instead of the
WordPress app, as we focus on improving the former. This change is
currently scoped to mobile web browsers, as we have not updated all the
CTAs displayed for non-mobile web browsers; we do not want to display
conflicting CTAs.

The button to email oneself an app install link is replaced with a link
to the Jetpack marketing site as it is not obvious the link would send
an email, it would require additional copy to explain. Instead, we
direct to the Jetpack marketing site containing additional details on
the mobile app.