Launchpad: Add early return to launchpad redirect statement (#67773)
The page client side router was rendering the home page instead of a loading screen when we were redirecting to launchpad.

This is because, previously, we inadvertently allowed client side routing to continue processing the /home/:siteId route, even after calling our redirect function. We now include an early return, which prevents homepage rendering and, instead, displays a loading screen.