Limit Global Styles: Show modal after opening the styles panel (#67748)
Displays a modal after opening the Global Styles panel in order to inform users that it's a paid feature. It has been implemented as part of the Editing Toolkit plugin, so we can target both Simple and Atomic sites (although Atomic sites are deliberately excluded at this point because of its lack of blog stickers support).