fix: Prevent display of an empty banner
Because the `HOME` section is only used when the relevant feature flag
is enabled, `getAppBannerData` is capable of returning empty title and
copy values. However, the `ALLOWED_SECTIONS` is a static assignment, so
this resulted in the app banner erroneously displaying with no title or
copy for the `HOME` section when the feature flag was disabled.