Update the auto-renew toggle to use a link on Purchase Details (#68105)
* Update the auto-renew toggle to use a link on Purchase Details

This component is used in several places. All other places should remain
intact.

The change will replace the Toggle component with an On/Off button in
the middle of the Auto-renew ON copy.

* There's no cursor and the tabbing doesn't work

Use a link button in order to make that work

* Rename the Modal heading

@DavidRothstein raised a concern about the flow not being immediately
clear with this heading. Rename it to make it more clear what it is