Use flexbox to vertically center text
For the domains section of the launchpad, selecting the domain to copy would highlight the height of the entire url box, not the height of the text.

This is because we were vertically centering text by matching line-height to height of the surrounding container. We now use an alternative method of vertical centering with flexbox to circumvent the problem entirely.