Ensure link text is correctly translated
It was mentioned here that splitting strings can be problematic: https://github.com/Automattic/wp-calypso/pull/68293#discussion_r981251709

As such, this commit uses the "translate" function's componentization feature to join the link in this paragraph up with the main text. It'll now be rendered as a full string, which will be easier for translators to handle.