Allow saved cards in checkout if saved cards are allowed by backend
The API endpoints (eg: the shopping-cart endpoint) return
`WPCOM_Billing_MoneyPress_Stored` if saved cards are allowed to be used
as a payment method.

For some reason `isPaymentMethodEnabled()` ignored that value and relied
on `WPCOM_Billing_Stripe_Payment_Method` which represents new cards.

Here we remove the unnecessary logic and instead make sure that we
normalized saved card slugs.