Jetpack Search: don't change jetpack_search to wpcom_search
Currently, when purchasing a Jetpack Search plan on wpcom
site (simple or atomic), we automatically change the product slug
from jetpack_search to wpcom_search,
and from jetpack_search_monthly to wpcom_search_monthly.

In this commit I'm removing that swap, as we now want to
use jetpack_search variant for both Jetpack and wpcom sites.