Agency Sign-Up: fix a bug in the company-details-form
A bug made the country fields appear even though the country list didn't be available. It happened because the logic was to show the country fields when the length of the country list was different from 0. The issue here is the country list will always have at least one item, the default item. This commit fixed this by showing the country fields only when the country list length is