Fix domain ownership verification flow redirectiion
A page redirection error occurs when domain ownership verification is
needed to connect a subdomain, and the subdomain already points to
WPCOM. In that scenario, at the end of the verification flow, the user
is redirected to the beginning of the "Use a domain I own" flow due to
a page slug error.

This commit redirects the user to the correct domain connection flow
step.