Jetpack Cloud: Update Jetpack app badge URL and styles (#69629)
* feat: Update Jetpack iOS App Store URL

This updated URL is now used for SEO purposes and used heavily in other
marketing materials.

* fix: Remove possibly troublesome app badge styles

It was not possible to prove these styles were the source of the issue,
but these styles may cause the link to intermittently fail in certain
context, e.g. use an iOS device with the Google Chrome browser.