Hide "Edit" and "Remove" buttons from the checkout line item when rendered in a Woo mobile webview (#69595)
When a Woo mobile app user is going through the flow to set up a new store, they need to purchase the wpcom plan for ecommerce. During this step, the composite checkout has "Edit" and "Remove" buttons on the plan line item, but if the user interacts with either of these, it will break the flow. This change employs user agent detection to conditionally hide these buttons when the checkout is rendered in a Woo mobile app webview.