Store: Modify thank you card for Woo mobile
When setting up a new store via the Woo mobile app, the user can create
a new wpcom account as part of the process if they don't have one yet,
and then they need to add the ecommerce plan to their new site (other
plans are not supported by the Woo mobile app). Once these steps are
complete, the `AtomicStoreThankYouCard` component is displayed. However,
when this renders in the Woo mobile app's webview, we don't want it to
show a message about verifying email (app creates a new account without
email verification) and we don't want it to provide a "Create your
store" button or any other CTA, as the mobile app will automatically
redirect the screen once everything is set up. This change uses
userAgent detection to conditionally modify the thankyou card component
accordingly in the Woo mobile app webview.